# src/spa/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Library Management System",
    "version": "V1"
  },
  "paths": {
...
    "/api/Checkout/card/{libraryCardId}": {
      "get": {
        "tags": [
          "Checkout"
        ],
        "parameters": [
          {
            "name": "libraryCardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "OrderBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortDirection",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SearchString",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutForListDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutForListDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutForListDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/LibraryCard": {
      "get": {
        "tags": [
          "LibraryCard"
        ],
        "parameters": [
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "OrderBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortDirection",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SearchString",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LibrarycardForListDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibrarycardForListDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibrarycardForListDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "LibraryCard"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryCardForCreationDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryCardForCreationDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryCardForCreationDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryCardForCreationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryCardForDetailedDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryCardForDetailedDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryCardForDetailedDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "LibraryCard"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryCardForUpdate"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryCardForUpdate"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryCardForUpdate"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryCardForUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryAssetForDetailedDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryAssetForDetailedDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryAssetForDetailedDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/LibraryCard/{cardId}": {
      "get": {
        "tags": [
          "LibraryCard"
        ],
        "operationId": "GetById",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryAssetForDetailedDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryAssetForDetailedDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryAssetForDetailedDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "LibraryCard"
        ],
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryAssetForDetailedDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryAssetForDetailedDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryAssetForDetailedDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/LibraryCard/cardnumber/{cardNumber}": {
      "get": {
        "tags": [
          "LibraryCard"
        ],
        "parameters": [
          {
            "name": "cardNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryAssetForDetailedDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryAssetForDetailedDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryAssetForDetailedDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Member/card/{cardId}": {
      "get": {
        "tags": [
          "Member"
        ],
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddressDto": {
        "required": [
          "city",
          "stateId",
          "street",
          "zipcode"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "street": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "stateId": {
            "type": "integer",
            "format": "int32"
          },
          "state": {
            "$ref": "#/components/schemas/StateDto"
          },
          "zipcode": {
            "maxLength": 5,
            "minLength": 5,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "AuthorDto": {
        "required": [
          "fullName"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "fullName": {
            "maxLength": 50,
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "maxLength": 250,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CategoryDto": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "maxLength": 250,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CheckoutForListDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "cardNumber": {
            "type": "string",
            "nullable": true
          },
          "checkoutDate": {
            "type": "string",
            "format": "date-time"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time"
          },
          "dateReturned": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "LibraryAssetForDetailedDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "photoUrl": {
            "type": "string",
            "nullable": true
          },
          "added": {
            "type": "string",
            "format": "date-time"
          },
          "numberOfCopies": {
            "type": "integer",
            "format": "int32"
          },
          "copiesAvailable": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "assetType": {
            "type": "string",
            "nullable": true
          },
          "authors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthorDto"
            },
            "nullable": true
          },
          "isbn": {
            "type": "string",
            "nullable": true
          },
          "deweyIndex": {
            "type": "string",
            "nullable": true
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "LibraryCardForCreationDto": {
        "required": [
          "address",
          "dateOfBirth",
          "firstName",
          "gender",
          "lastName",
          "phoneNumber"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "maxLength": 25,
            "minLength": 1,
            "type": "string"
          },
          "lastName": {
            "maxLength": 25,
            "minLength": 1,
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "phoneNumber": {
            "minLength": 1,
            "type": "string",
            "format": "tel"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date-time"
          },
          "address": {
            "$ref": "#/components/schemas/AddressDto"
          },
          "gender": {
            "$ref": "#/components/schemas/MemberGenderDto"
          }
        },
        "additionalProperties": false
      },
      "LibraryCardForDetailedDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "cardNumber": {
            "type": "string",
            "nullable": true
          },
          "fees": {
            "type": "number",
            "format": "double"
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "address": {
            "$ref": "#/components/schemas/AddressDto"
          },
          "addressId": {
            "type": "integer",
            "format": "int32"
          },
          "gender": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "libraryCardPhoto": {
            "type": "string",
            "nullable": true
          },
          "age": {
            "type": "integer",
            "format": "int32"
          },
          "photoUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LibrarycardForListDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "cardNumber": {
            "type": "string",
            "nullable": true
          },
          "gender": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "libraryCardPhoto": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LibraryCardForUpdate": {
        "required": [
          "email",
          "firstName",
          "lastName",
          "phoneNumber"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "maxLength": 25,
            "minLength": 2,
            "type": "string"
          },
          "lastName": {
            "maxLength": 25,
            "minLength": 2,
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phoneNumber": {
            "type": "string",
            "format": "tel"
          },
          "address": {
            "$ref": "#/components/schemas/AddressDto"
          },
          "gender": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "MemberGenderDto": {
        "enum": [
          "Male",
          "Female",
          "Other"
        ],
        "type": "string"
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
...
      "StateDto": {
        "required": [
          "abbreviations",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "abbreviations": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}